requirements.txt
# Required python packages
numpy
matplotlib
ray
scipy
Pillow
pytest
imageio
